Upgrade to 64-bit libraries
Since EasyOpenCV does not build its JNI library in 64-bit, we have to
compile our own (libs/arm64-v8a/libEasyOpenCV.so). Additionally,
EasyOpenCV uses its custom shared library loader and expects a 32-bit
OpenCV library in /sdcard/FIRST (see DynamicOpenCvNativeLibLoader.java
in https://github.com/OpenFTC/OpenCV-Repackaged). This loading will
fail, so we have to add the OpenCV library as a dependency of
libEasyOpenCV.so. The Android dynamic linker only allows us to load
libraries from /data and /mnt/expand, so we placed the OpenCV library in
/data and chmod it so that it is readable by others. Note that
libEasyOpenCV.so has to be compiled with DT_RUNPATH set to /data so that
the dynamic linker can find libopencv_java4.so.